% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_fct.R
\name{load_data}
\alias{load_data}
\title{Download panTHERIA or Wildfinder dataset}
\usage{
load_data(overwrite = FALSE, name_base_data, filename)
}
\arguments{
\item{overwrite}{a \code{logical}. If \code{TRUE}, the file will be downloaded again
and the previous version will be replaced.}

\item{name_base_data}{can be PanTHERIA or Wildfinder given the data we want to download}

\item{filename}{is the name of the datafile to download}
}
\value{
No return value.
}
\description{
This function downloads the PanTHERIA or Wildfinder datasets (text file) hosted on the
GitHub repository \url{https://github.com/rdatatoolbox/datarepo/}. The file
won't be downloaded if already exists locally (except if \code{overwrite = TRUE}).
}
